+++ modules/cloudwatch-logs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## cloudwatch-logs
 
+### 0.0.6 / 14.11.2023
+* [Change] Use 'for_each' meta-argument instead of 'count' to avoid unnessary changes when the log_groups list changes. 
+
 ### 0.0.5 / 4.10.2023
 * [bug_fix] Add secret_manager_enabled variable to allow the use of lambda-SecretLayer in the same run with this module 
 
@@ -16,3 +19,4 @@
 
 ### 0.0.1 / 3.8.2023
 * [Update] Add support to use a private link with coralogix - add subnet_id and security_group_id variable to connect the lambda to vpc.
+