+++ modules/resource-metadata/CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ## resource-metadata
 
-### 0.0.6 / 27.11.2023
+### 0.0.7 / 12.11.2023
 * [Change] Add an option to specify the retention time of the CloudWatch log group that is created by the lambda
+
+### 0.0.6 / 11.12.2023
+* [Update] Add filtering of Lambda functions.
 
 ### 0.0.5 / 4.10.2023
 * [bug_fix] Add secret_manager_enabled variable to allow the use of lambda-SecretLayer in the same run with this module